+ Add: Base tab v.1.0
- Modified: ActorTab.cs improvement and integration with BaseTab.cs
